requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.5
django-shortcuts==1.6
djangorestframework==3.12.4
drf-spectacular==0.17.2
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
